Implement the "hero" layer of landing page
had some issue with image not taking 50%
of the width of the screen. Was able to
solve this my wrapping the image with a
div, then giving the div 50% of the width
then giving the image width of 100%. I had
to make the image heigth auto or else the
image will look distorted.

Overall, there is too much blank space on
the hero layer due to the image heigth being
taller than the total heigth of the hero texts
heigth.